aider: Added AWS as the global provider with the region set to "us-west-2" in terragrunt.hcl.
# Context:
USER: with this codebase where can we set an overall provider to be used for all deployments?
ASSISTANT: In a Terraform codebase, the provider is typically defined in the `main.tf` file within each module. However, if you want to set a global provider for all deployments, you can do so in the root `terragrunt.hcl` file.

Please note that this file is currently read-only. If you want to make changes to it, you need to change its permissions to read-write.
USER: I added these *read-write* files: terragrunt.hcl
ASSISTANT: Great! Now, to set a global provider for all deployments, you can add a `provider` block in the `terragrunt.hcl` file. Here's an example of how to set AWS as the provider:

```hcl
terragrunt.hcl
<<<<<<< ORIGINAL

=======
provider "aws" {
  region = "us-west-2"
}
>>>>>>> UPDATED
```

Please replace `"us-west-2"` with the AWS region you want to use for your deployments. If you're using a different provider, replace `"aws"` and the `region` attribute with the appropriate values for your provider.