aider: Set Azure as the global provider.
# Context:
USER: update the provider to use Azure, this entire codebase is for my azure lab.
ASSISTANT: Sure, to set Azure as the global provider, you need to modify the `provider` block in the `terragrunt.hcl` file. Here's how you can do it:

```hcl
terragrunt.hcl
<<<<<<< ORIGINAL
provider "aws" {
  region = "us-west-2"
}
=======
provider "azurerm" {
  features {}
}
>>>>>>> UPDATED
```

This sets Azure as the provider. The `features {}` block is required by the Azure provider, but it can be left empty. If you need to specify a specific version of the Azure provider or configure it further, you can do so within this `provider` block.